go uin64 to uin32

+++ go/cert_srv/reiss_tasks.go
@@ -84,7 +84,7 @@ func (s *SelfIssuer) createLeafCert(leaf *cert.Certificate, config *conf.Conf) e
 	}
 	chain := &cert.Chain{Leaf: leaf.Copy(), Issuer: issCrt}
 	chain.Leaf.Version += 1
-	chain.Leaf.IssuingTime = uint64(time.Now().Unix())
+	chain.Leaf.IssuingTime = uint32(time.Now().Unix())
 	chain.Leaf.CanIssue = false
 	chain.Leaf.ExpirationTime = chain.Leaf.IssuingTime + cert.DefaultLeafCertValidity
 	if chain.Issuer.ExpirationTime < chain.Leaf.ExpirationTime {
@@ -115,7 +115,7 @@ func (s *SelfIssuer) createIssuerCert(config *conf.Conf) error {
 		return err
 	}
 	crt.Version += 1
-	crt.IssuingTime = uint64(time.Now().Unix())
+	crt.IssuingTime = uint32(time.Now().Unix())
 	crt.CanIssue = true
 	crt.ExpirationTime = crt.IssuingTime + cert.DefaultIssuerCertValidity
 	coreAS, err := s.getCoreASEntry(config)
@@ -222,7 +222,7 @@ func (r *ReissRequester) Run() {
 // currently active certificate chain.
 func (r *ReissRequester) sendReq(chain *cert.Chain, config *conf.Conf) error {
 	c := chain.Leaf.Copy()
-	c.IssuingTime = uint64(time.Now().Unix())
+	c.IssuingTime = uint32(time.Now().Unix())
 	c.ExpirationTime = c.IssuingTime + (chain.Leaf.ExpirationTime - chain.Leaf.IssuingTime)
 	c.Version += 1
 	if err := c.Sign(config.GetSigningKey(), chain.Leaf.SignAlgorithm); err != nil {
